pyproject.toml
[project]
name = "copaw"
dynamic = ["version"]
description = "CoPaw is a **personal assistant** that runs in your own environment. It talks to you over multiple channels (DingTalk, Feishu, QQ, Discord, iMessage, etc.) and runs scheduled tasks according to your configuration. **What it can do is driven by Skills — the possibilities are open-ended.** Built-in skills include cron, PDF/Office handling, news digest, file reading, and more; you can add custom skills. All data and tasks run on your machine; no third-party hosting."
readme = "README.md"
requires-python = ">=3.10,<3.14"
dependencies = [
    "agentscope==1.0.16.dev0",
    "agentscope-runtime==1.1.0",
    "httpx>=0.27.0",
    "discord-py>=2.3",
    "dingtalk-stream>=0.24.3",
    "uvicorn>=0.40.0",
    "apscheduler>=3.11.2,<4",
    "playwright>=1.49.0",
    "questionary>=2.1.1",
    "mss>=9.0.0",
    "reme-ai==0.3.0.5",
    "transformers>=4.30.0",
    "python-dotenv>=1.0.0",
    "python-socks>=2.5.3",
    "onnxruntime<1.24",
    "lark-oapi>=1.5.3",
    "python-telegram-bot>=20.0",
    "twilio>=9.10.2",
]

[tool.setuptools.dynamic]
version = {attr = "copaw.__version__.__version__"}

[tool.setuptools]
packages = { find = { where = ["src"] } }
include-package-data = true

[tool.setuptools.package-data]
# Dockerfile copies console build output to src/copaw/console/ (no dist subdir).
"copaw" = [
    "console/**",
    "agents/md_files/**",
    "agents/skills/**",
    "tokenizer/**",
]

[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project.scripts]
copaw = "copaw.cli.main:cli"

[project.optional-dependencies]
dev = [
    "pytest>=8.3.5",
    "pytest-asyncio>=0.23.0",
    "pre-commit>=4.2.0",
    "pytest-cov>=6.2.1",
]
full = [
    "huggingface_hub>=0.20.0",
    "llama-cpp-python>=0.3.0",
    "mlx-lm>=0.10.0",
    "ollama>=0.6.1",
]
local = [
    "huggingface_hub>=0.20.0",
]
llamacpp = [
    "copaw[local]",
    "llama-cpp-python>=0.3.0",
]
mlx = [
    "copaw[local]",
    "mlx-lm>=0.10.0",
]
ollama = [
    "ollama>=0.6.1",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]
